logdet() is logdet_spd()
