+++ docs/log_determinants.md
@@ -23,7 +23,7 @@ Estimate log-determinants as such:
 >>> sample_fun = montecarlo.normal(shape=(6,))
 >>> matvec = lambda x: a.T @ (a @ x) + x
 >>> order = 3
->>> logdet = slq.logdet(matvec, order, key=key, sample_fun=sample_fun)
+>>> logdet = slq.logdet_spd(matvec, order, key=key, sample_fun=sample_fun)
 >>> print(jnp.round(logdet))
 3.0
 >>> # for comparison:
